WR: (pdf, endpoints) unification of code divergence
squash

squash
